requirements.txt
fastapi
pydantic
uvicorn
pymongo[srv]
python-dotenv
streamlit
requests